phase-2/backend/requirements.txt
fastapi[all]>=0.104.1
sqlmodel>=0.0.8
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.24.0